+++ index.js
@@ -12,29 +12,33 @@ function fetchJSON (url) {
     })
 }
 
-function price (fsym, tsyms, tryConversion) {
+function price (fsym, tsyms, options) {
+  options = options || {}
   let url = `${baseUrl}price?fsym=${fsym}&tsyms=${tsyms}`
-  if (tryConversion === false) url += '&tryConversion=false'
+  if (options.tryConversion === false) url += '&tryConversion=false'
   return fetchJSON(url)
 }
 
-function priceMulti (fsyms, tsyms, tryConversion) {
+function priceMulti (fsyms, tsyms, options) {
+  options = options || {}
   let url = `${baseUrl}pricemulti?fsyms=${fsyms}&tsyms=${tsyms}`
-  if (tryConversion === false) url += '&tryConversion=false'
+  if (options.tryConversion === false) url += '&tryConversion=false'
   return fetchJSON(url)
 }
 
-function priceFull (fsyms, tsyms, tryConversion) {
+function priceFull (fsyms, tsyms, options) {
+  options = options || {}
   let url = `${baseUrl}pricemultifull?fsyms=${fsyms}&tsyms=${tsyms}`
-  if (tryConversion === false) url += '&tryConversion=false'
+  if (options.tryConversion === false) url += '&tryConversion=false'
   // We want the RAW data, not the DISPLAY data:
   return fetchJSON(url).then(result => result.RAW)
 }
 
-function priceHistorical (fsym, tsyms, time, tryConversion) {
+function priceHistorical (fsym, tsyms, time, options) {
+  options = options || {}
   time = dateToTimestamp(time)
   let url = `${baseUrl}pricehistorical?fsym=${fsym}&tsyms=${tsyms}&ts=${time}`
-  if (tryConversion === false) url += '&tryConversion=false'
+  if (options.tryConversion === false) url += '&tryConversion=false'
   // The API returns json with an extra layer of nesting, so remove it
   return fetchJSON(url).then(result => result[fsym])
 }